requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
boto3==1.34.28
botocore==1.34.28
certifi==2023.11.17
cffi==1.16.0
channels==3.0.4
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.7
django-allauth==0.58.2
django-braces==1.15.0
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-storages==1.14.2
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jmespath==1.0.1
jwcrypto==1.5.0
msgpack==1.0.7
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyotp==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
razorpay==1.4.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
soupsieve==2.5
sqlparse==0.4.4
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.7
whitenoise==6.6.0
wrapt==1.16.0
zope.interface==6.1
